Cleaned up reimport of wayland-protocol sources
The existing files were a bit of a mess, and it was not clear how to
update the protocol files from upstream.

This CL clearly identifies where each protocol file comes from, and
there are now hopefully simple instructions we can follow to perform
updates.

As part of cleaning it up, I've uprevved the protocols to the latest
versions from the upstream sources.

I've also created a Soong "wayland_protocol_codegen" build rule
(implemented by a corresponding .go file) to handle the code generation
steps automatically, including the filename transformations we need.

Bug: None
Test: master-arc-wm-dev builds, Wayland unit tests pass
Change-Id: I0623ce981281c464c821c5146930916b43906575